webtester/requirements-expand-task.txt
google-generativeai
python-dotenv
pytest  # For running tests